feat(notes): enhanced settings and refactored primary view
- enhanced icons to use internal icons rather than custom icons
- updated how the storage location setting is handled
- extension will prompt for reload when storage location change is detected
- settings icon will force user to select a storage location if no storage location, otherwise it will take user to extension settings
- updated dependencies and basic tests
