+++ src/types/mkdirp.d.ts
@@ -1,0 +1,4 @@
+declare module 'mkdirp' {
+    function mkdirp(dir: string): Promise<string>;
+    export = mkdirp;
+}
