+++ src/types/rimraf.d.ts
@@ -1,0 +1,4 @@
+declare module 'rimraf' {
+    function rimraf(path: string, callback: (error: Error | null) => void): void;
+    export = rimraf;
+}
