+++ tsconfig.json
@@ -8,7 +8,12 @@
 		],
 		"sourceMap": true,
 		"rootDir": "src",
-		"strict": true   /* enable all strict type-checking options */
+		"strict": true, /* enable all strict type-checking options */
+		"esModuleInterop": true, /* Enables emit interoperability between CommonJS and ES Modules */
+		"typeRoots": [
+			"./node_modules/@types",
+			"./src/types"
+		],
 		/* Additional Checks */
 		// "noImplicitReturns": true, /* Report error when not all code paths in function return a value. */
 		// "noFallthroughCasesInSwitch": true, /* Report errors for fallthrough cases in switch statement. */
